Initial crypto implementation
Includes WIP implementations for `getEntropy`, `secretKeyToMnemonic`,
`mnemonicToMasterSeed`, `hashPassword`, `getSKeysFromPassword`, and
`regenerateMnemonic`. Did not yet touch AEAD stuff and signing stuff.

I think this is a good starting point because everything up to
`regenerateMnemonic` can be used in an integration test to ensure we can
generate a mnemonic, process it, and then use that to regenerate the
mnemonic.

The folder structure is the way it is to allow for tree-shaking later
on. We will configure each directory to be a separate entry point /
output, so we can do:

```javascript
import getEntropy from 'nex-auth-protocol/getEntropy'`
```

...and not incur loading for the entire package.

Problems:
* Not a lot of good library support for crypto functions. What's the
best way forwards?
* Bad type support for existing crypto functions, but not a big problem.
* How to effectively unit test crypto functions? Asserting known input
=> known output doesn't seem very robust.

* Refactor: Clean up existing functions

* scrypt-async => scrypt-js
* implement getSKeysFromPassword with HKDF
* switch all salts to userID
* switch mnemonics to be represented by array of strings vs. space-delimited string
* add util to normalize string and convert to buffer
* add some specs

* Remove randombytes in favor of crypto (to be polyfilled)
* General cleanup
* Add declaration file for futoin-hkdf

crypto is node only